+++ SmallBeachFarm/Framework/FishType.cs
@@ -1,0 +1,12 @@
+namespace Pathoschild.Stardew.SmallBeachFarm.Framework
+{
+    /// <summary>The fish types that can be caught.</summary>
+    internal enum FishType
+    {
+        /// <summary>Fish that can be caught in the river.</summary>
+        River,
+
+        /// <summary>Fish that can be caught in the ocean.</summary>
+        Ocean
+    }
+}