+++ SmallBeachFarm/ModEntry.cs
@@ -35,6 +35,9 @@ namespace Pathoschild.Stardew.SmallBeachFarm
         /// <summary>The mod configuration.</summary>
         private ModConfig Config;
 
+        /// <summary>The mod's hardcoded data.</summary>
+        private ModData Data;
+
         /// <summary>The minimum value to consider non-transparent.</summary>
         /// <remarks>On Linux/Mac, fully transparent pixels may have an alpha up to 4 for some reason.</remarks>
         private const byte MinOpacity = 5;
@@ -52,6 +55,7 @@ namespace Pathoschild.Stardew.SmallBeachFarm
         {
             // read config
             this.Config = helper.ReadConfig<ModConfig>();
+            this.Data = helper.Data.ReadJsonFile<ModData>("assets/data.json") ?? new ModData();
 
             // hook events
             helper.Events.Player.Warped += this.OnWarped;
@@ -65,7 +69,7 @@ namespace Pathoschild.Stardew.SmallBeachFarm
                 addCampfire: this.Config.AddCampfire,
                 useBeachMusic: this.Config.UseBeachMusic,
                 isSmallBeachFarm: location => this.IsSmallBeachFarm(location, out _),
-                isOceanTile: this.IsOceanTile
+                getFishType: this.GetFishType
             );
         }
 
@@ -193,7 +197,7 @@ namespace Pathoschild.Stardew.SmallBeachFarm
             GameLocation beach = Game1.getLocationFromName("Beach");
             foreach (CrabPot pot in farm.objects.Values.OfType<CrabPot>())
             {
-                if (this.IsOceanTile(farm, (int)pot.TileLocation.X, (int)pot.TileLocation.Y))
+                if (this.GetFishType(farm, (int)pot.TileLocation.X, (int)pot.TileLocation.Y) == FishType.Ocean)
                     pot.DayUpdate(beach);
             }
         }
@@ -229,23 +233,35 @@ namespace Pathoschild.Stardew.SmallBeachFarm
             return false;
         }
 
-        /// <summary>Get whether a given position is ocean water.</summary>
+        /// <summary>Get the fish that should be available from the given tile.</summary>
         /// <param name="farm">The farm instance.</param>
         /// <param name="x">The tile X position.</param>
         /// <param name="y">The tile Y position.</param>
-        private bool IsOceanTile(Farm farm, int x, int y)
+        private FishType GetFishType(Farm farm, int x, int y)
         {
-            // check water property
+            // not water
+            // This should never happen since it's only called when catching a fish, but just in
+            // case fallback to the default farm logic.
             if (farm.doesTileHaveProperty(x, y, "Water", "Back") == null)
-                return false;
+                return FishType.River;
 
-            // check for beach tilesheet
+            // mixed fish area
+            if (this.Data.MixedFishAreas.Any(p => p.Contains(x, y)))
+            {
+                return Game1.random.Next(2) == 1
+                    ? FishType.Ocean
+                    : FishType.River;
+            }
+
+            // ocean or river
             string tilesheetId = farm.map
                 ?.GetLayer("Back")
                 ?.PickTile(new Location(x * Game1.tileSize, y * Game1.tileSize), Game1.viewport.Size)
                 ?.TileSheet
                 ?.Id;
-            return tilesheetId == "zbeach" || tilesheetId == "zbeach_farm";
+            return tilesheetId == "zbeach" || tilesheetId == "zbeach_farm"
+                ? FishType.Ocean
+                : FishType.River;
         }
 
         /// <summary>Get whether the player shouldn't be able to access a given position.</summary>